backend/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
cryptography==2.9
Django==3.0.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.9
isodate==0.6.0
pycparser==2.20
pyparsing==2.4.6
pytz==2019.3
rdflib==4.2.2
requests==2.23.0
six==1.14.0
SPARQLWrapper==1.8.5
sqlparse==0.3.1
urllib3==1.25.8
sentry-sdk
django-crontab
whitenoise
psycopg2==2.7.7
dj-database-url==0.4.2